### 2021-10-12 #471 @msweetlove
OmicsMetaData: an R-package for interoperable and re-usable biodiversity 'omics (meta)data
Submitting Author Name: Maxime Sweetlove
Submitting Author Github Handle: @msweetlove
Other Package Authors Github handles: @ymgan, @Antonarctica
Repository:  https://github.com/biodiversity-aq/OmicsMetaData
Version submitted: 0.0.1
Submission type: Standard

Version accepted: TBD

---

```
Package: OmicsMetaData
Title: OmicsMetaData: an R-package for interoperable and re-usable biodiversity 'omics (meta)data.
Version: 0.0.1
	person("Maxime",
		"Sweetlove",
		email = "msweetlove@naturalsciences.be")
		OmicsMetaData: tools to re-use biodiversity 'omics datasets or standardize them to make them globally interoperable
License: GNU General Public License v3 (GLP-3.0) https://www.gnu.org/licenses/gpl-3.0.en.html
    worrms (>=0.4.2),
    stringr (>=1.4.0),
    Orcs (>=1.2.1),
    xml2 (>=1.3.2),
    reshape2 (>=1.4.4)
    tidyr (>=1.1.3),
    RecordLinkage(>=0.4-12.1),
    mapview (>=2.10.0),
    rgbif (>=3.6.0),
    testthat (>=2.3.2),
    roxygen2 (>=7.1.1),
    devtools (>=2.3.0)
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under:

data retrieval: The OmicsMetaData package allows users to download nucleotide sequences from INSDC alongside any associated metadata linked to the sequences, based on a BioProject identification number.

data munging: The OmicsMetaData package provides tools to format biodiversity 'omics metadata following the widely used data standards MIxS (for 'Omics data) and DarwinCore (for biodiversity data). Standardizing datasets makes them much more easily interoperable between researchers and institutions and across time, and makes it easier for users to archive the metadata alongside the sequences on the INSDC databases.

Scientists that work with biodiversity 'omics datasets on a daily basis, and have a need to standardize the data to exchange it between colleagues or archive it online after the end of a project, or if they want to expand their dataset with online openly available sequence data.

At present I am not aware of any such packages.
