+++ DESCRIPTION
@@ -12,19 +12,19 @@ License: GNU General Public License v3 (GLP-3.0) https://www.gnu.org/licenses/gp
 Encoding: UTF-8
 LazyData: true
 Imports:
-    worrms (>=0.4.2),
-    stringr (>=1.4.0),
-    Orcs (>=1.2.1),
-    xml2 (>=1.3.2),
-    reshape2 (>=1.4.4)
+    worrms (>= 0.4.2),
+    stringr (>= 1.4.0),
+    Orcs (>= 1.2.1),
+    xml2 (>= 1.3.2),
+    reshape2 (>= 1.4.4)
 Depends:
-    tidyr (>=1.1.3),
-    RecordLinkage(>=0.4-12.1),
-    mapview (>=2.10.0),
-    rgbif (>=3.6.0),
+    tidyr (>= 1.1.3),
+    RecordLinkage(>= 0.4-12.1),
+    mapview (>= 2.10.0),
+    rgbif (>= 3.6.0),
     R (>= 4.0.1)
 RoxygenNote: 7.1.1
 Suggests: 
-    testthat (>=2.3.2),
-    roxygen2 (>=7.1.1),
-    devtools (>=2.3.0)
+    testthat (>= 2.3.2),
+    roxygen2 (>= 7.1.1),
+    devtools (>= 2.3.0)
