+++ DESCRIPTION
@@ -19,7 +19,7 @@ Imports:
     reshape2 (>= 1.4.4)
 Depends:
     tidyr (>= 1.1.3),
-    RecordLinkage(>= 0.4-12.1),
+    RecordLinkage (>= 0.4-12.1),
     mapview (>= 2.10.0),
     rgbif (>= 3.6.0),
     R (>= 4.0.1)
