### 2021-10-15 88355a6
bug fixes after R CMD
