+++ R/Get_SequenceData_INSDC.R
@@ -20,6 +20,7 @@
 #' @seealso get.sample.attributes.INSDC
 #' @return if get.BioProject.metadata.INSDC(just.names=FALSE) (default) a data.frame with n rows an m columns is returned, n being the number of samples in the BioProject, and m being the number of variables found. If get.BioProject.metadata.INSDC(just.names=TRUE), a character vector of length n is returned with the sample numbers ("Run numbers", SRR numbers)
 #' @references Sayers, E. (2009) The E-utilities In-Depth: Parameters, Syntax and More, https://www.ncbi.nlm.nih.gov/books/NBK25499/
+#' @importFrom utils download.file read.table read.csv
 #' @examples 
 #' \donttest{
 #' get.BioProject.metadata.INSDC(BioPrjct="PRJNA303951", just.names=FALSE)
@@ -53,13 +54,14 @@ get.BioProject.metadata.INSDC <- function(BioPrjct=NA, just.names=FALSE){
 #' @family downloading data functions
 #' @description Downloads all sample attributes (that is, additional environmental or other associated data) from INSDC. Note: requires a user-specified API-key to acces the INSDC databases. see https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/ to generate an API-key
 #' @param sampleID a list. a list of one ore more SRA sample IDs (that is "Run" numbers). This argument can be left blank if input is provided for the BioPrjct argument (see further)
-#' @param apiKey a character string. A personal API-key to the acces the NCBI databases, and required to use the Entrez Programming Utilities (E-utilities). An API-key (API stands for application programming interface) is a unique identifier used to authenticate a user. A personal API-key can easily be generated at requested at https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
+#' @param apiKey a character string. A personal API-key to the access the NCBI databases, and required to use the Entrez Programming Utilities (E-utilities). An API-key (API stands for application programming interface) is a unique identifier used to authenticate a user. A personal API-key can easily be generated at requested at https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
 #' @param BioPrjct a character string or a vector with character strings. Providing the associated BioProject numbers helps to give more understandable error messages. Alternativey, if the sampleID argument is empty, all sample IDs from the given BioProject numbers will be given to the sampleID argument.
 #' @details Each sequence data sample ("Run") typically has additional measurements or metadata associated with it. However, these are difficult to find, and cannot be downloaded together with the nucleotide sequence data. The get.sample.attributes.INSDC function will fetch the all the metadata of the samples given to the sampleID (or BioPrjct) argument from the INSDC. To do this, get.sample.attributes.INSDC will use the Entrez Programming Utilities (E-utilities) from NCBI, and will access the databases with the API-key provided by the user. see https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/ for more info on getting an API-key.
 #' @return a dataframe with the data found
+#' @importFrom xml2 read_xml xml_find_all xml_attr xml_text as_list
 #' @examples 
 #' \donttest{
-#' get.BioProject.metadata.INSDC(BioPrjct="PRJNA303951", apiKey="YouPersonalAPIKey")
+#' get.sample.attributes.INSDC(BioPrjct="PRJNA303951", apiKey="YouPersonalAPIKey")
 #' }
 #' @export
 get.sample.attributes.INSDC <- function(sampleID=NA, apiKey=NA, BioPrjct=NA){
[…]
