+++ R/DataQC_Main_DwCgeneral.R
@@ -24,7 +24,7 @@
 #' @details DarwinCore is the biodiversity data standard develloped by TDWG, and is used by the Global Biodiversity Information Facility (GBIF). This function performs a basic and user-supervised quality control. This includes cheking all variables terms adhere to the DarwinCore vocabulary, listing other variables in an eMoF file or in the dynamicProperties field, and checking for obvious errors in the content of the data (typos, different NA values,...)
 #' @return a dataframe that is formatted as either: an event or occurrence
 #' @examples 
-#' \donttest{
+#' \dontrun{
 #' test_event <- data.frame(eventID=c("sample1", "sample2"),
 #'                          eventDate=c("2021-09-27", "2021-09-28"),
 #'                          decimalLatitude=c("54.7", "33"),
