### 2021-10-20 da7e85a
additional unit tests
