+++ tests/testthat/test_Get_SequenceData_INSDC.R
@@ -1,26 +1,47 @@
 #==============================================================
 # Author Maxime Sweetlove
-# lisence CC-0
+# license CC-0
 # Part of the POLA3R website (successor or mARS.biodiversity.aq)
 # version 1.0 (2021-09-23)
-# file encdong UTF-8
+# file encding UTF-8
 #
 #==============================================================
 library(OmicsMetaData)
 library(testthat)
 
 test_that("get.BioProject.metadata.INSDC works", {
-  expect_s3_class(get.BioProject.metadata.INSDC(BioPrjct="PRJNA303951", just.names=FALSE), "data.frame")
+  expect_s3_class(get.BioProject.metadata.INSDC(BioPrjct="PRJNA303951", just.names=FALSE),
+                  "data.frame")
+  expect_equal(class(get.BioProject.metadata.INSDC(BioPrjct="PRJNA303951", just.names=TRUE)),
+               "character")
   expect_error(get.BioProject.metadata.INSDC(BioPrjct="nonsens_DezeNaamBestaatNiet_mooiweervandaag", just.names=FALSE))
 })
 
 
 # get.sample.attributes.INSDC can't be fully tested due to needing a personal API key
-test_that("download.sequences.INSDC works", {
-  expect_message(download.sequences.INSDC(BioPrj = "PRJNA303951", destination.path = NA, apiKey=NA,
-                                          unzip = FALSE, keep.metadata = FALSE, download.sequences = FALSE),
-                 c("Getting the metadata ...", "Processing BioProject PRJNA303951...", 
-                   "50 samples (Runs)...", "Finished processing", "No files were downloaded"))
+test_that("get.BioProject.metadata.INSDC works", {
+  expect_error(get.sample.attributes.INSDC(sampleID="SRR3087847", apiKey=NA))
+  expect_error(get.sample.attributes.INSDC(BioPrjct="PRJNA303951", apiKey=NA))
 })
+
+
+test_that("download.sequences.INSDC works", {
+  expect_message(download.sequences.INSDC(BioPrj = "PRJNA303951", destination.path = NA, 
+                                          apiKey=NA, unzip = FALSE, keep.metadata = FALSE, 
+                                          download.sequences = FALSE))
+  expect_error(download.sequences.INSDC(BioPrj = "PRJNA303951", destination.path = NA, 
+                                          apiKey="test", unzip = FALSE, keep.metadata = TRUE, 
+                                          download.sequences = FALSE))
+  expect_error(download.sequences.INSDC(BioPrj = "nonsens_DezeNaamBestaatNiet_mooiweervandaag", destination.path = NA, 
+                                          apiKey=NA, unzip = FALSE, keep.metadata = TRUE, 
+                                          download.sequences = FALSE))
+  expect_error(download.sequences.INSDC(BioPrj = "nonsens_DezeNaamBestaatNiet_mooiweervandaag", destination.path = NA, 
+                                        apiKey=NA, unzip = TRUE, keep.metadata = TRUE, 
+                                        download.sequences = FALSE))
+  
+})
+# clean up
+closeAllConnections()
+
 
 
