### 2021-10-22 faa7dd2
expand test coverage
