+++ tests/testthat/test_DataQC_DwC.R
@@ -9,14 +9,13 @@
 library(OmicsMetaData)
 library(testthat)
 
-test_event <- data.frame(eventID=c("sample1", "sample2"),
-                            eventDate=c("2021-09-27", "2021-09-28"),
-                            decimalLatitude=c("54.7", "33"),
-                            decimalLongitude=c("88.9", "-48.4"),
-                            row.names=c("sample1", "sample2"))
-
-
-test_that("dataQC.DwC_general works for event data", {
+test_that("dataQC.DwC_general works for event data, no user input", {
+  test_event <- data.frame(eventID=c("sample1", "sample2"),
+                           eventDate=c("2021-09-27", "2021-09-28"),
+                           decimalLatitude=c("54.7", "33"),
+                           decimalLongitude=c("88.9", "-48.4"),
+                           row.names=c("sample1", "sample2"))
+  
   expect_s3_class(dataQC.DwC_general(dataset=test_event, DwC.type = "event", ask.input = FALSE,
                                      complete.data=TRUE), "data.frame")
   expect_s3_class(dataQC.DwC_general(dataset=test_event, DwC.type = "event", ask.input = FALSE,
@@ -29,7 +28,67 @@ test_that("dataQC.DwC_general works for event data", {
                                   complete.data=TRUE), test_event)
 })
 
+test_that("dataQC.DwC_general works for event data, with user input", {
+  test_event_1 <- data.frame(eventid=c("sample1", "sample2"),
+                           eventDate=c("2021-09-27", "2021-09-28"),
+                           decimalLatitude=c("54.7", "33"),
+                           decimalLongitude=c("88.9", "-48.4"),
+                           row.names=c("sample1", "sample2"))
+  test_event_2 <- data.frame(staalnaam=c("sample1", "sample2"),
+                           datum=c("2021-09-27", "2021-09-28"),
+                           latitude=c("54.7", "33"),
+                           longitude=c("88.9", "-48.4"),
+                           row.names=c("sample1", "sample2"))
+  test_event_3 <- data.frame(eventid=c("sample1", "sample2"),
+                             eventids=c("sample1", "sample2"),
+                             eventdate=c("2021-09-27", "2021-09-28"),
+                             extravar1=c("88.9", "-48.4"),
+                             extravar2=c("sample1", "sample2"))
+  test_event_1OUT <- data.frame(eventID=c("sample1", "sample2"),
+                              eventDate=c("2021-09-27", "2021-09-28"),
+                              decimalLatitude=c("54.7", "33"),
+                              decimalLongitude=c("88.9", "-48.4"),
+                              row.names=c("sample1", "sample2"))
+  test_event_2OUT <- data.frame(staalnaam=c("sample1", "sample2"),
+                                datum=c("2021-09-27", "2021-09-28"),
+                                decimalLatitude=c("54.7", "33"),
+                                decimalLongitude=c("88.9", "-48.4"),
+                                dynamicProperties=c("{\"staalnaam\":sample1}", "{\"staalnaam\":sample2}"),
+                                eventRemark = c("datum:2021-09-27", "datum:2021-09-28"),
+                                eventID = c("perfix_1", "perfix_2"),
+                                row.names=c("sample1", "sample2"))
+  test_event_3OUT <- data.frame(eventID=c("sample1", "sample2"),
+                                eventids=c("sample1", "sample2"),
+                                eventDate=c("2021-09-27", "2021-09-28"),
+                                extravar2=c("sample1", "sample2"),
+                                fieldNotes=c("extravar2:sample1", "extravar2:sample2"))
[…]
