+++ tests/testthat/test_DataQC_Utils.R
@@ -1,30 +1,40 @@
 #==============================================================
 # Author Maxime Sweetlove
-# lisence CC-0
 # Part of the POLA3R website (successor or mARS.biodiversity.aq)
 # version 1.0 (2021-09-23)
-# file encdong UTF-8
+# file encding UTF-8
 #
 #==============================================================
 library(OmicsMetaData)
 library(testthat)
 
-dates <- c("2020-09-23", "2020", "16 Jan. 2020", "November 1998", "12/01/1999")
-latitudes <- c(23, 45, -56.44, "47.5", "-88° 4\' 5\"")
-longitudes <- c(23, 45, -56.44, "47.5", "-88° 4\' 5\"")
-sampleNames <- paste("sample", 1:5, sep="_")
-test_metadata <- data.frame(sample_name=sampleNames,
-                           collection_date=dates,
-                           latitude=latitudes,
-                           longitude=longitudes)
+
 
 test_that("dataQC.dateCheck delivers on dates it can handle", {
+  dates <- c("2020-09-23", "2020", "16 Jan. 2020", "November 1998", "12/01/1999")
+  latitudes <- c(23, 45, -56.44, "47.5", "-88° 4\' 5\"")
+  longitudes <- c(23, 45, -56.44, "47.5", "-88° 4\' 5\"")
+  sampleNames <- paste("sample", 1:5, sep="_")
+  test_metadata <- data.frame(sample_name=sampleNames,
+                              collection_date=dates,
+                              latitude=latitudes,
+                              longitude=longitudes)
+  
   expect_equal(dataQC.dateCheck(dataset=test_metadata, date.colnames=c("collection_date")),
                list(values=c("2020-09-23","2020","2020-01-16","1998-11","1999-01-12"), 
                     warningmessages=NULL))
 })
 
 test_that("dataQC.LatitudeLongitudeCheck delivers on values it can handle", {
+  dates <- c("2020-09-23", "2020", "16 Jan. 2020", "November 1998", "12/01/1999")
+  latitudes <- c(23, 45, -56.44, "47.5", "-88° 4\' 5\"")
+  longitudes <- c(23, 45, -56.44, "47.5", "-88° 4\' 5\"")
+  sampleNames <- paste("sample", 1:5, sep="_")
+  test_metadata <- data.frame(sample_name=sampleNames,
+                              collection_date=dates,
+                              latitude=latitudes,
+                              longitude=longitudes)
+  
   expect_equal(dataQC.LatitudeLongitudeCheck(dataset=test_metadata, 
                                              latlon.colnames=list(c("lat_lon"),c("latitude"), c("longitude"))),
                list(values=c("23 23","45 45","-56.44 -56.44","47.5 47.5","-87.9319 -87.9319"), 
@@ -32,10 +42,36 @@ test_that("dataQC.LatitudeLongitudeCheck delivers on values it can handle", {
 })
 
 test_that("dataQC.guess.env_package.from.data works fine", {
+  dates <- c("2020-09-23", "2020", "16 Jan. 2020", "November 1998", "12/01/1999")
+  latitudes <- c(23, 45, -56.44, "47.5", "-88° 4\' 5\"")
+  longitudes <- c(23, 45, -56.44, "47.5", "-88° 4\' 5\"")
+  sampleNames <- paste("sample", 1:5, sep="_")
+  test_metadata <- data.frame(sample_name=sampleNames,
+                              collection_date=dates,
+                              latitude=latitudes,
+                              longitude=longitudes)
+  
   expect_equal(dataQC.guess.env_package.from.data(dataset=test_metadata, pckge.colnames=c("env_package", "ScientificName")),
                list(values=NULL, 
[…]
