+++ tests/testthat/test_DataFormat_Utils.R
@@ -1,115 +1,149 @@
 #==============================================================
 # Author Maxime Sweetlove
-# lisence CC-0
 # Part of the POLA3R website (successor or mARS.biodiversity.aq)
 # version 1.0 (2023-09-23)
-# file encdong UTF-8
+# file encding UTF-8
 #
 #==============================================================
 library(OmicsMetaData)
 library(testthat)
 
-#outputs are the correct format (including dimensions and components)
-#sample input produces the correct sample output
-
-# generate test data
-sampleNames <- c("sample_1", "sample_2")
-
-# a mock extended measurement or fact (eMoF) table
-test_emof <- data.frame(eventID = c(rep(c("sample_1", "sample_2"), 2)),
-                        measurementType = factor(c(rep("var1", 2), rep("var2", 2))), 
-                        measurementValue = c(1:4),
-                        measurementUnit = c(rep("unit1", 2), rep("unit2", 2)))
-# a mock wide formated table
-test_wideTab <- list(data = data.frame(var1=c(1,2), var2=c(3,4), row.names=sampleNames),
-                     units = c(var1="unit1", var2="unit2"),
-                     method=NA)
-
-# a mock MIxS.metadata object
-test_MIxS <- new("MIxS.metadata",
-                 data   = data.frame(var1=c(1,2), var2=c(3,4), eventID=sampleNames, row.names=sampleNames),
-                 section = c(var1="section1", var2="section1", eventID="miscellaneous"),
-                 units      = c(var1="unit1", var2="unit2", eventID="alphanumeric"),
-                 env_package = "water",
-                 type = "versatile",
-                 QC = TRUE)
+test_that("combine.data.frame gives correct output", {
+  # generate test data
+  df01 <- data.frame(var1=1:4, var2=2:5, row.names=paste("r", 1:4, sep=""))
+  df02 <- data.frame(var1=5:8, var3=1:4, row.names=paste("t", 1:4, sep=""))
+  df03 <- data.frame(var1=5:8, var3=1:4, row.names=c(paste("r", 1:3, sep=""), "q1"))
+  df12 <- data.frame(var1=1:8, 
+                     var2=c(2:5, rep(NA, 4)), 
+                     var3=c(rep(NA, 4), 1:4),
+                     row.names=c(paste("r", 1:4, sep=""), paste("t", 1:4, sep="")))
+  
+  
+  expect_equal(combine.data.frame(df01, df02, fill=NA, merge.cols=TRUE, 
+                                  original_rowName.col=TRUE, merge.rows="df1"), df12)
+  expect_s3_class(combine.data.frame(df01, df02, fill="-", merge.cols=FALSE, 
+                                  original_rowName.col=TRUE, merge.rows="df1"), "data.frame")
+  expect_s3_class(combine.data.frame(df01, df03, fill="-", merge.cols=TRUE, 
+                                     original_rowName.col=TRUE, merge.rows=NA), "data.frame")
+  
+  })
 
 # test the eMoF.to.wideTable function
 test_that("eMoF.to.wideTable gives correct output", {
-  expect_equal(eMoF.to.wideTable(test_emof), test_wideTab)
-  expect_error(eMoF.to.wideTable(test_emof[2:4]))
+  sampleNames <- c("sample_1", "sample_2")
+  test_emof_1 <- data.frame(eventID = c(rep(c("sample_1", "sample_2"), 2)),
+                          measurementType = factor(c(rep("var1", 2), rep("var2", 2))), 
+                          measurementValue = c(1:4),
+                          measurementUnit = c(rep("unit1", 2), rep("unit2", 2)))
+  test_emof_2 <- data.frame(occurrenceID = c(rep(c("sp1", "sp2"), 2)),
[…]
