+++ R/DataFormat_Utils.R
@@ -1,15 +1,14 @@
 #==============================================================
 # Author Maxime Sweetlove
-# lisence CC-0
 # Part of the POLA3R website (successor or mARS.biodiversity.aq)
 # version 1.0 (2019-09-20)
-# file encdong UTF-8
+# file encding UTF-8
 #
 #==============================================================
 
 #' convert a DarwinCore extended Measurement or Fact (eMoF) file to a wide tabular format
 #' @family formating functions
-#' @author Maxime Sweetlove CC-0 2020
+#' @author Maxime Sweetlove
 #' @description converts a DarwinCore extended Measurement Or Fact (eMOF) file, which is has a "long" file format into a wide tabular format
 #' @usage eMoF.to.wideTable(dataset)
 #' @param dataset a dataframe of the eMoF file
@@ -37,7 +36,7 @@ eMoF.to.wideTable <- function(dataset){
   if("eventID" %in% colnames(dataset)){
     dataLong<-dataset[,colnames(dataset) %in% c("eventID", "measurementType", "measurementValue")]
     dataWide <- tidyr::spread(data=dataLong, "eventID", "measurementValue")
-  }else if("occurrenceID" %in% colnames(dataLong)){
+  }else if("occurrenceID" %in% colnames(dataset)){
     dataLong<-dataset[,colnames(dataset) %in% c("occurrenceID", "measurementType", "measurementValue")]
     dataWide <- tidyr::spread(dataLong, "occurrenceID", "measurementValue")
   }else{
@@ -79,7 +78,7 @@ eMoF.to.wideTable <- function(dataset){
 
 #' convert dataframe to a DarwinCore extended Measurement or Fact (eMoF) file
 #' @family formating functions
-#' @author Maxime Sweetlove CC-0 2020
+#' @author Maxime Sweetlove
 #' @description converts a dataframe to a DarwinCore extended Measurement Or Fact (eMOF) file
 #' @usage wideTable.to.eMoF(metadata.object, variables)
 #' @param metadata.object a MIxS.metadata object
@@ -152,7 +151,7 @@ wideTable.to.eMoF <- function(metadata.object, variables=NA){
 }
 
 #' merge dataframes
-#' @author Maxime Sweetlove CC-0 2019
+#' @author Maxime Sweetlove
 #' @family formating functions
 #' @description combine.data.frame merges two dataframes, completing the rows and columns that are not shared by the dataframes.
 #' @usage combine.data.frame(df1, df2, fill=NA, merge.cols=TRUE, original_rowName.col=TRUE, merge.rows="df1")
@@ -263,7 +262,7 @@ combine.data.frame <- function(df1, df2, fill=NA,
       df_x <- data.frame(matrix(nrow = nrow(df2)-length(shared_rows), ncol = ncol(df11), data=fill))
       colnames(df_x) <- colnames(df11)
       rownames(df_x) <- setdiff(rownames(df2), shared_rows)
-      df_notshared_1 <- cbind(df_x, df2[setdiff(rownames(df2), shared_rows),])
+      df_notshared_2 <- cbind(df_x, df2[setdiff(rownames(df2), shared_rows),])
     }else{
       df_notshared_2 <- data.frame()
     }
@@ -277,7 +276,7 @@ combine.data.frame <- function(df1, df2, fill=NA,
 }
 
 #' merge MIxS.metadata objects
-#' @author Maxime Sweetlove CC-0 2019
+#' @author Maxime Sweetlove
 #' @family formating functions
 #' @description combine.data combines two MIxS.metadata objects into one, merging all common variables (columns as default).
 #' @usage combine.data(d1, d2, fill=NA, variables.as.cols=TRUE)
[…]
