### 2021-10-25 33c754c
addressing cmd notes
