### 2021-10-25 0d8728a
added example for commonTax.to.NCBI.TaxID.Rd
