+++ R/General_Utils.R
@@ -6,7 +6,7 @@
 # Author Maxime Sweetlove
 # Part of the POLA3R website (successor or mARS.biodiversity.aq)
 # version 1.0 (2019-09-20)
-# file encding UTF-8
+# file encoding UTF-8
 #
 #==============================================================
 # Tools for general use
@@ -19,6 +19,11 @@
 #' @usage multi.warnings(message_text, warningmessages)
 #' @param message_text a character string. The error message to be added
 #' @param warningmessages a vector with one ore more character strings. The previous error messages.
+#' @examples 
+#' \donttest{
+#' warningmessages <- c()
+#' multi.warnings(warningmessages, "new warning")
+#' }
 #' @return a character string with unique warning messages
 multi.warnings <- function(message_text, warningmessages){
   if(! message_text %in% warningmessages){
@@ -34,6 +39,10 @@ multi.warnings <- function(message_text, warningmessages){
 #' @param taxon character or character vector. The taxon names to be converted to NCBI tax IDs.
 #' @param fill.unknown character. The string to return when a taxon was not found in the list with common taxa. Default is NCBI:txid12908 (the ID for "unknown sequences"), other options include NA or ""
 #' @details This function makes use of a limited internal library, not all the taxIDs are present
+#' @examples 
+#' \donttest{
+#' commonTax.to.NCBI.TaxID("soil metagenome", fill.unknown="12908")
+#' }
 #' @return a character or character vector with the matching taxon IDs, with fill.unknown for those not found.
 commonTax.to.NCBI.TaxID<-function(taxon, fill.unknown="12908"){
 
@@ -65,6 +74,10 @@ commonTax.to.NCBI.TaxID<-function(taxon, fill.unknown="12908"){
 #' @author Maxime Sweetlove
 #' @description get the ENA variant a MIxS term
 #' @param variable character a MIxS term.
+#' @examples 
+#' \donttest{
+#' get.ENAName("target_gene")
+#' }
 #' @return a character
 get.ENAName <- function(variable){
   ENAName <- as.character(TermsLib[TermsLib$name==variable,]$name_variant_ENA)
