### 2021-11-03 f655705
corrected examples of not exported functions
