### 2021-11-08 1e9490e
goodPractises: consistency in assigning
