### 2021-12-23 #471 @ginberg
---

README
- contains a typo: 'lisence' instead of license
- styling: some (sub)titles are not showing up as subtitles since there is no space between the hashtags and the title

Check:
- devtools is listed twice in Suggests section in DESCRIPTION

Vignettes:
- styling: some (sub)titles are not showing up as subtitles since there is no space between the hashtags and the title

Automated tests:
- The overall test coverage is good (78.91%). There are 20 warnings when running all tests, could you fix these?

Packaging guidelines:
- rOpenSci recommends a package name in lowercase in their [guidelines](https://devguide.ropensci.org/building.html), could you change that?
