### 2022-02-06 #471 @cpalmer718
Hi @msweetlove @jooolia please see below for my review. I've added commentary at the bottom to explain some of it. I'm trying to return this early because I wasn't able to clearly evaluate parts of the package on my system, so we may need to iterate a bit.

---

  - Windows
    - Based on the ropensci specs, I'm assuming this is intended to work on Windows, but I guess I don't know for certain. In any case, my primary test system happened to be Windows 10, but the package example under "Brief demonstration usage" does not work on Windows due to hard-coded use of wget. This also seems to break the testthat tests. It seems like this package was developed on OSX; was this tested on Windows or Linux? I recommend replacing "wget" with "auto" to allow R to figure it out.
    - Assigning encoding as "UTF-8" without conversion fails on Windows 10 (degree symbol is declared invalid UTF-8, tests fail). this is obviously very likely platform dependent. passing character vector through appropriate sanitizer `enc2utf8` fixes the problem.
    - A vignette example in "Retrieving online data" fails on Windows with the error `Error in download.file(ftp_url, destfile = file.path(destination.path,  : scheme not supported in URL 'ftp.sra.ebi.ac.uk/vol1/fastq/SRR298/004/SRR2980684/SRR2980684_1.fastq.gz'.` Apparently the "auto" setting is causing Windows to choose a download method that doesn't support ftp. This is seemingly resolved by setting the method parameter in `download.file` to "libcurl" as it is available on my system. I'd recommend going through the `download.file` documentation and coming up with a logic chain that hopefully guarantees the download will succeed for any system (which may involve requiring non-R system dependencies), or otherwise exposing the method parameter to the user so they can override the system's choice without editing the library's source code.
  - OSX/Linux
    - After having problems with Windows, I tried installing on OSX and Linux, and had different issues. I tried installing in clean conda environments and ran up against a number of system dependencies that failed. Can you please try out clean installations on those systems and document system dependencies that may be required (e.g. `libudunits2`, `gdal`, etc.)?
  - README.md has various issues with markdown syntax that are interfering with legibility: code blocks formatted as R, spaces after '#' characters in headers, etc.
  - The "Settings" page for NCBI doesn't have a clear link once you click to their site; it might be helpful to provide clearer instructions about how to get to the right page.
  - The way the API key is ingested in the package, it seems like it's encouraging users to paste it into scripts, which represents a kind of security vulnerability. It would be nice to instruct users how they might better secure their API key for use in R (for example placing it in .Renviron or using a more formal handling mechanism.
  - Similar to README.md, working on improving the markdown syntax and presentation could really help the legibility of the vignettes. They're hard to get through at the moment.
  - Please add `author: "{name}"` tags to YAML headers of vignette .Rmd files
  - Be sure to change links to other vignettes from within the vignettes to point to .html files instead of .Rmd source files. It would also be helpful to link a table of contents at the top of each vignette, or something like that, given the number of vignettes.
  - `workingDir` is defined in `Metadata_standardization` but then not used, and in the final statement `write.MIxS` function emits relative to working directory
  - In `Retrieving online data`, the example in block `download_sequences` emits the warning `the metadata will be retruned to the Console If you did assign the output of this function to an R-object (using "<-"): better abort and restart now`; since the result is assigned to a variable in the vignette example, this is really confusing as a new user.
  - It seems that no vignette documents anything about DarwinCore content in the library, unless I've missed it somehow.
  - I think, as a general comment, it's hard to get from the vignettes what I'm really looking for as a new user, which is an end-to-end example of what the intended use case is for the package. I think this is in part due to how the vignettes are fragmented, and also that for most of the vignettes the code is not executable as-is. Though it's hard, this experience might be substantially improved by consolidating the vignettes into end-to-end walkthroughs of how you envision the user directly interacting with a dataset, with every step of data management and adjustment provided with a runnable example, and maybe some tables showing the expected data frame contents or similar. This would also possibly provide an opportunity to include DarwinCore content in the vignettes.
#### General repo comments:
  - The license status of the project is unclear. Top level README.md says GPL3 (please be sure to correct typo). However, source files and codedata.json all say CC-0. Please either:
    - harmonize this license selection across all files, and for good measure add license declarations in the header of all source files; or
    - explain in top-level README.md the license status of the project so it's clear
  - R history and data hidden files, and Mac tracker files (`.DS_Store`), should be removed from the repo. The appropriate file patterns should be added to `.gitignore` at top level to prevent their addition to the repo.
  - The convention for default branch naming has changed (see [here](https://github.com/github/renaming)). Please consider renaming the default branch from `master` to `main` or `default` as preferred.
  - `R/sysdata.rda` has something called "MarsLib" in it; is that expected to be there? It seems like several data structures are being loaded directly from saved data files, which are provided as jagged data frames in the package namespace. It works, but it's somewhat unwieldy to work with. It would be nice to have some of the prestock library content available within the library as YAML files or something. Regardless, it really seems like the jagged data frames should be rather lists of vectors, so that when the user accesses one of the shorter vectors, they don't accidentally end up using the many empty string entries at the end of the vector.
#### Style/packaging:
  - The package would really benefit from passes through `styler` and `lintr`. In particular:
    - the cyclomatic complexity is really high for some functions, which makes it very challenging to actually test them.
    - variable and function naming is a bit messy. the combination of camel and dotted snake and all-caps and other styles in the same names at times makes it very difficult to read. the ropensci docs recommend snake_case.
  - Some of the automated checks recommend avoiding `sapply` and some other similar things, and I think that would be good. In particular, there are some statements (e.g. `unlist(unname(sapply(paste))))` that I think could be made much more straightforward by using, for example, paste's vectorized behavior. This doesn't seem to cause any bugs that I've noticed, but it makes it difficult to read and maintain, and depending on the size of your input datasets, some of the for loops might cause some issues. I don't think it really will cause performance degradation at realistic dataset sizes, but it would just make things a lot cleaner and consistent with packaging best practices.
#### Tests:
  - The tests didn't catch some of the issues I had with the package on Windows, so they really could use expansion.
  - Some of the tests technically cover the functions but don't then, for example, check the format of the file emitted by the tested function. I strongly encourage increasing test coverage and thoroughness if you can.

Thank you for inviting me to review! In particular with the Windows system issues, I'd really enjoy taking another look once things are working on the platform.
