### 2022-03-18 #471 @jooolia
Dear @ginberg  and @cpalmer718,
Thank you very much for your reviews. We appreciate the time and care you put into them. The author has switched jobs and cannot continue with the development of this package and unfortunately there is no-one at his former workplace who can carry on with the work so the package development will be stopped. This means that we will close this review now. Please let me know how many hours you each spent reviewing (you can estimate if you can't exactly remember as it was quite some time ago) and we can log this in our reviewer database.

Thanks again for your help and input,
Julia
